shoppley.com/shoppley/requirements.txt
--find-links=http://pypi.pinaxproject.com
--find-links=http://pypi2.pinaxproject.com

# Get all requirements formerly known as "external_libs" explicitly
# hashlib
python-dateutil==1.5
docutils
elementtree

gdata
Pygments
python-openid
python-twitter
#python-yadis==1.1.0
textile
vobject
BeautifulSoup
ybrowserauth
Markdown
diff-match-patch
geopy
pytz
creole

django-ajax-validation
django-announcements
django-atomformat
django-avatar
django-bookmarks
django-sentry
django-debug-toolbar
django-email-confirmation
django-extensions
django-flag
django-friends
django-gravatar
django-locations
django-mailer
django-messages
django-microblogging
django-notification
django-oembed
django-openid
django-pagination
django-robots
django-swaps
django-sorting
django-tagging
django-tagging-ext
django-threadedcomments
django-timezones
django-uni-form
django-voting
django-wikiapp
django-photologue
django-attachments
django-markup
template_utils
django-autofixture
django-groups
sorl-thumbnail
django-wakawaka
